Simplify cursor usage, and avoid keeping track of the cursor history
NB: I believe cursors do not change as new reviews are written, so it is
more consistent than the start_offset used by Steam API before. Cursors
pinpoint specific reviews, which could be useful to resume the download
of reviews if we are missing old reviews. To download new reviews, there
is no need to keep track of cursors.